requirements.txt
fastapi
uvicorn
pandas
scikit-learn
tensorflow
motor
shap
python-dotenv
